requirements.txt
scikit_build
pytest
pyfakefs
pymongo
cerberus
plumbum
validators
jira
jenkinsapi
python-jenkins
python-debian
mockupdb
configparser
luigi
python-gnupg
pyone
zmq
GitPython==2.1.10
gitdb2==2.0.6
pygments
pygerrit
flup
u-msgpack-python
retrying
lmdb
#rpm Conflicts with python3-rpm
pyicu
pebble
pyyaml
mock
PyJWT==1.7.1
Flask==0.12.2
flask-classful==0.14.1
Flask-PyMongo==0.5.1
flask-jwt-extended>=3.12.1
Flask-Misaka
memoization
paramiko
fabric
raven
